Add notifications for problematic sound directories and checks for unusable files
Upon specifying in the module settings a directory path that isn't accessible or has no sound files usable by Foundry, the user will be notified of this via a ui notification warning as well as a console warning.
Files unusable as audio clips by Foundry will no longer accidentally be selected to be played as a critical sound-clip.